Updated admin side, new shortcode, added tdd_pb_get_bars() function
* Moved the logic that actually displays a plugin out of the short code
function and into it's own tdd_pb_get_bars() function which will allow
us to make this a widget very easily in the future.

* Also because of that code shift-around, the actual progress bar is
displayed in wp-admin -- hopefully making it a touch easier to preview
what the things look like. Next step is to make the quick-edit box
actually display the percentage and color. (the bar is also displayed
on the edit screen)

* In addition to the shortcode [progress], the plugin will now accept
[tdd_pb] although this is more to appease people that have been mucking
around in the code ("tdd_pb" prefixes functions and variables)