Quick Edit and Fixing CSS for FF
* Given the major change to the CSS, the javascript to make things
animate also had to change pretty extensively. One side effect is that
the animations on a page will eventually slow to a standstill at the
rate of about 1-2 pixels every 15 seconds or so. The animation is still
a work in progress...
* Major bug fixes for the CSS to show up in, well, pretty much
everything but Chrome. Didn't realize I was using an unsupported style.
* Small CSS fix for the "Strawberry" color - now shows up correctly
* Quick Edit mode on the "Manage Progress Bars" page now has the
ability to change percentage and bar color. No need to specifically
edit the progress bar now.