Additional display options, options for calculated percentage, and a filter hook! -- ready for 0.4
* Progress bars can now be defined using a percentage or two numbers (
x of y ) and let the plugin do the calculation
* The text display on the bar itself can be nothing, the percentage,
the two numbers, or both. The global setting to turn this on and off
still exists.
* Changed the checkbox on the admin page to make it clear it is a
global "off" switch for bar text, not just percentage.
* Added a filter hook for the percentage being shown. This provides the
ability to hook in your own method of calculating percentages. The hook
name is `tdd_pb_calculated_percentage`